Fixes for use behind reverse proxy in production system
Implemented ReverseProxyViewHandler to be used only in systems located
behind a reverse proxy (usually production). The view handler
essentially converts absolute referenced resource and url locations and
converts them to relative while also stripping out the context path
(/gotrack in this case). They are commented out for now but should be
uncommented when merging to master.