Adds contextual help to Enrichment Tracking, See Issue #11
Reasoning for the complicated update on Run Enrichment button: Updating
the buttons that display the help context on the tab titles breaks the
linkage to their overlayPanel. Cannot group the buttons and overlayPanel
together as that messes with DOM widths. The option I went with was to
exclude the tabView from the update and then manually enable the tabs as
a callback.