Fix for previous workaround on enrichment plot disabled markers
Previously we disabled markers on the point level because disabling it
at the series level removed series symbols from the legend. We now solve
this by forcing non-zero stroke-width at the css level in the legend and
adding a nominal linewidth to the series which is not enough to see but
enough to force HC to draw the line in the legend item.