Major refactor of backend for new SQL schema and logic
The logic of a unique symbol per Gene no longer holds. Instead, we have
a single unqiue accession per Gene. This means that methods which
returned a Gene via a species and symbol no longer make sense. In
addition, this means we must change the way Gene views and searches
work as (species, symbol) is no longer enough to fully describe a Gene.
Now, the accession is used to fully describe a Gene.

Bundled into this commit are also some rolling cleanup refactorings,
such as having method use Species instead of species_id.