Stability score now shows significance volatility
Coefficient of variation now based on p-value
signifiance cutoff instead of mean. This gives
a sense of how volatile it is WRT its signifiance.

Stability colour scale is now absolute instead of
quantile based and is a bit more lenient.

Stability column infinity and NaN replaced with
No Change and No Data, respectively.

Related to issue #36.