Replace overlayPanel with tooltips for help icons
overlayPanels cause many headaches when attempting
to use them for anything but the most generic cases.

Specifically, this change fixes display bugs in FF as
well as an interaction with datatable exporting
that causes column headers to show a meaningless
class toString when an overlay panel exists in the
header facet. Rather than hack a workaround, tooltips
were decided to be a simpler and more maintainable
approach site-wide.

We use jquery tooltip instead of tipsy purely for the
functionality of being able to attach the tooltip item
selector to document thus allowing dynamic creation of
elements with help icons (re: dynamically loaded
row expansion).